requirements.txt
asgiref==3.3.1
Django==3.1.7
mssql-django==1.0a5
pyodbc==4.0.30
python-dotenv==0.15.0
pytz==2021.1
sqlparse==0.4.1
